+++ apps/vscode-extension/src/utils/getActionTemplate.ts
@@ -101,6 +101,29 @@ export const getActionTemplate = (selectedAction: QuickPickItem): any => {
     };
   }
 
+  if (action === Action.ShowQR) {
+    return {
+      action: Action.ShowQR,
+      url: '',
+      topText: '',
+      title: '',
+      description: '',
+      logo: '',
+    };
+  }
+
+  if (action === Action.HideQR) {
+    return {
+      action: Action.HideQR,
+    };
+  }
+
+  if (action === Action.ClosePreview) {
+    return {
+      action: Action.ClosePreview,
+    };
+  }
+
   /**
    * Text actions
    */