+++ apps/webviews/src/components/preview/index.ts
@@ -4,6 +4,7 @@ export * from './LaserPointer';
 export * from './Markdown';
 export * from './MarkdownPreview';
 export * from './ProjectorIcon';
+export * from './QRPreview';
 export * from './SlideControl';
 export * from './SlideControls';
 export * from './WhiteboardIcon';